+++ src/controller/ReportWebviewController.ts
@@ -1,8 +1,9 @@
 import vscode, { Disposable } from 'vscode'
-import { ProfilerConfig, ProjectReport, UnifiedPath } from '@oaklean/profiler-core'
+import { UnifiedPath } from '@oaklean/profiler-core'
 
 import { Container } from '../container'
 import WorkspaceUtils from '../helper/WorkspaceUtils'
+import { ProjectReportHelper } from '../helper/ProjectReportHelper'
 
 export class ReportWebviewController implements Disposable {
 	private container: Container
@@ -19,14 +20,13 @@ export class ReportWebviewController implements Disposable {
 		try {
 			vscode.window.showInformationMessage('Opening JSON editor...')
 			const inputPath = new UnifiedPath(filePath)
-			const config =  WorkspaceUtils.autoResolveConfigFromPath(inputPath.dirName())
+			const config = WorkspaceUtils.autoResolveConfigFromPath(inputPath.dirName())
 			if (config === undefined) {
 				vscode.window.showErrorMessage(`Could not find a profiler config at ${inputPath.toPlatformString()}`)
 				return
 			}
-			const report = ProjectReport.loadFromFile(inputPath, 'bin', config)
-			if (report === undefined) {
-				vscode.window.showErrorMessage(`Could not find a profiler report at ${inputPath.toPlatformString()}`)
+			const report = ProjectReportHelper.loadReport(inputPath, config)
+			if (report === null) {
 				return
 			}
 			const formattedJson = JSON.stringify(report, null, 2)