+++ src/controller/TextEditorController.ts
@@ -22,13 +22,14 @@ export default class TextEditorController implements Disposable {
 			this.container.eventHandler.onProgramStructureTreeChange(this.programStructureTreeChanged.bind(this)),
 			this.container.eventHandler.onReportLoaded(this.reportLoaded.bind(this)),
 			this.container.eventHandler.onSelectedSensorValueTypeChange(this.selectedSensorValueTypeChanged.bind(this)),
-			this.container.eventHandler.onToggleLineAnnotationsChange(this.toggleLineAnnotationsChange.bind(this))
+			this.container.eventHandler.onToggleLineAnnotationsChange(this.toggleLineAnnotationsChange.bind(this)),
+			vscode.window.onDidChangeActiveColorTheme(() => {
+				this.refresh()
+			})
 		)
-
 		this.textDecorations = []
 		this._enableLineAnnotations = this.container.storage.getWorkspace('enableLineAnnotations', true) as boolean
 		this._sensorValueRepresentation = this.container.storage.getWorkspace('sensorValueRepresentation') as SensorValueRepresentation
-
 	}
 
 	toggleLineAnnotationsChange(event: ToggleLineAnnotationsChangeEvent) {